requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
pyyaml==6.0.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dateutil==2.8.2
apscheduler==3.10.4
